Updating the code for generating MoM
